Com-726 - remove unecessary aggregate
